Add NC methods
This commit introduces several NC methods, both in the Classifier module
and in the v1 module. This commit includes methods to delete all
children of a node group, various attribute extractors, and the ability
to PUT and POST to update node groups.

Stylistically, it also removes all logic from the v1 module; this allows
for cleaner separation between the api calls and any transforms you
might want to implement.