(maint) use reachable_name for farady connection
Previous to this commit, the farady connection method was just using
`@dashboard`. However, beaker returns the node_name by default. This
works so long as the dashboard host is in your `/etc/hosts` file.
This commit changes it to use `dashboard.reachable_name` - which
defaults to IP address then hostname and is what beaker uses to
establish an ssh connection.