Set classifier path for update_classes endpoint
Previous to this commit, the `update_classes` endpoint in the classifier
API was not calling `set_classifier_path` resulting in the code going to
`https://dashboard/v1/update_classes` instead of the api url.
This commit fixes it by adding the `set_classifier_path` to the method.
This should probably be fixed more broadly so each method does not need
to call that method.