Disable ssl verification when using a Unix::Host
Previous to this commit, when passed a Unix::Host (a beaker object) to
connect to, scooter would reach to and talk to the API via the IP
address. However this will fail since the URL we are talking to (the
ip) is not on the cert name.
This commit disables SSL CA verification for now until a better fix can
be found.