Create new beaker method for ec2 public ip
Beaker and ec2 don't play nice with getting the public ip. It is only
set during initial provision and can be over-ridden. If you also have to
run the script several times, or are using an existing set of nodes for
testing, beaker has no way to get the ec2 instanes public ip address,
mostly because the box it self does not expose it anywhere.  According
to the docs, you can curl the below to get it
http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-instance-addressing.html