(QA-1925) Change signin to not autoacquire xcsrf
The console used to not have a console view permission, allowing any
authenticated entity view access; This allowed any user to acquire an
xcsrf token after successful login.

The signin method still tries to acquire the XCSRF token, but catches
failures if the dispatcher does not have view access to the console.