Add more RBAC helpers
This adds methods to generate local users and roles. Previously, the
logic to create a local user was embedded into the v1 module, which was
not the correct location for it. The method has been moved to the Rbac
module and renamed generate_local_user.

While this is a breaking change, none of the tests actually used that
method, and so it can be safely erased.