(maint) fix bug with attach_ds_to_rbac
A bug existed in the implementation of this method; if no ldapdispatcher
was provided, the local var settings would not be set and you would get
an uninitialized error for settings. This change fixes that issue so you
can supply an options hash with any set of directory_service parameters
you want.