(QA-1937) Create base httpdispatcher
Prior to this commit, the only real dispatcher available was the
consoledispatcher. The consoledispatcher had much of the base methods
defined in itself, which needed to be pulled out and put into another
class that all other dispatcher classes should be subclassing. This PR
creates a base httpdispatcher, making it more straightfoward to
subclass the httpdispatcher class and build different dispatching
objects.