(QA-1943) Add method for ou's created during old init
A previous commit moved the organizational unit creation from the init
to the base method to generate a test environment of users and groups
for RBAC; this broke any implementation not using that method to
generate a test environment.

This doesn't add that back, as putting this action in the init was a
mistake; it does, however, give the user the ability to create the ou's
independently, and set the instance variables accordingly.