(PE-12084) Ensure URIs are subclassed from URI::HTTPS
The http-cookie library used by faraday cookie-jar relies on the ancient
text of http-cookie, which has not received an update for over two
years. That cookie library erroneously demanded that all URIs must be
subclassed from URI::HTTPS, not accepting a URI::HTTP object with a
scheme set to https. This change hardcodes all the httpdispatchers to
use URI::HTTPS objects for the url_prefix. This should have zero effect
on other middleware, as URI::HTTPS is just a subclass of URI::HTTP.