(PE-15456) Optionally wait for update_classes to finish
Previously the update_classes method would trigger an update to the
classes known to the classifier, and immediately return. However, the
update-classes endpoint is asynchronous, which means that tests calling
this method that depend on the classes being updated have no guarantees
about it being ready.
This commit updates the method to use a newly added method that gets the
last time that the classes were updated from the last-class-update
endpoint. This way update_classes can wait for the last-class-update to
change to indicate that the update is finished. To preserve backward
compatibility, the waiting is optional and the new parameter defaults
to false.