(BOLT-731) Add scooter endpoints for pe-bolt-server
This adds a new http dispatcher for pe-bolt-server, and the first endpoint
`ssh/run_task`.  I opted to include all methods directly in the
`PEBoltServerDispatcher` class rather than have a separate `PEBolt::V1` class
for the public methods because it seems much simpler and less confusing for new
comers. This does break the pattern used here, but I'm willing to sacrifice
consistency for simplicity. Open to discussion!