(PE-37639) fetch login page prior to login
In order to simulate the login flow accurately, this adds a request
to the login page prior to posting a login request. The cookie is
extracted from the login page response (via Set-Cookie), and if it
is present, is included as a cookie in the POST to the login request
page.

Tests are updated to add a stub for the login page request prior
to the post.